testing[IRGOFilter]: make slight adjustments to tests
Make slight adjustments to go-filter tests:
- Move the go_library fixture to python/tests/unit. Don't make the default assumption that the go library file exists for integration tests. Some of the integration tests call assert_valid_envoy_config which runs a validation check on the generated xDS inside the base envoy docker container. Because the xDS was generated under the assumption the go library file existed it would cause the validate command to fail due to the file actually not existing inside the container. Making it so that it was injected would require building an object file of the go library and incorporate that into the integration tests which is fairly complicated at this time. So we'll omit it for the time being especially because these integration tests are not testing the go filter specifically and test the go filter with additional integration tests.
- Don't use post_error when the go library file is not found. This will cause it to not count as an ir error but it will still log an error level message. This is so that unrelated kat tests pass.
