+++ python/tests/unit/conftest.py
@@ -1,0 +1,10 @@
+from unittest.mock import patch
+
+import pytest
+
+
+@pytest.fixture(autouse=True)
+def go_library():
+    with patch("ambassador.ir.irgofilter.go_library_exists") as go_library_exists:
+        go_library_exists.return_value = True
+        yield go_library_exists
